+++ cloud/disk_manager/internal/pkg/clients/nfs/testing/filesystem_model.go
@@ -158,7 +158,7 @@ func (f *FileSystemModel) Create() {
 		},
 	)
 	for _, child := range f.root.Children {
-		f.CreateNodes(uint64(nfs.RootNodeID), child)
+		f.CreateNodes(nfs.RootNodeID, child)
 	}
 }
 
@@ -228,7 +228,7 @@ func (f *FileSystemModel) ListNodesRecursively(parentNodeID uint64) []nfs.Node {
 }
 
 func (f *FileSystemModel) ListAllNodesRecursively() []nfs.Node {
-	return f.ListNodesRecursively(uint64(nfs.RootNodeID))
+	return f.ListNodesRecursively(nfs.RootNodeID)
 }
 
 func (f *FileSystemModel) SetSession(session nfs.Session) {
